+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [?NEXT?] - ????-??-??
 
+- Added `theme` props to `LayoutRenderer`.
+- Added some more props in `LayoutRenderer`.
+
+## [v0.0.11] - 2018-05-07
+
 - Fix component path when name is `index`.
 
 ## [v0.0.10] - 2018-05-01